Make AuditableRepository more configurable with respect to owner identity.
Allow AuditableRepository to to be more flexible in which value get stored in CreatedBy and UpdatedBy, instead of forcing it to always use _principal.Identity.Name. This is useful in situations where the necessary auditable information may be some other unique identifier such as an SSO ID, email address, etc.